**Description** 1. Updated the readme for setting up local dev environment. 2. Set up sql Create Table script 3. Refactored Backend.
