+++ frontend/vite.config.js
@@ -3,5 +3,11 @@ import react from '@vitejs/plugin-react'
 
 // https://vitejs.dev/config/
 export default defineConfig({
-  plugins: [react()]
+  plugins: [react()],
+  server: {
+    headers: {
+      // Use restrict-properties instead of same-origin
+      'Cross-Origin-Opener-Policy': 'unsafe-none',
+    }
+  }
 })
